refactor(tests): share one compiler harness between the two probe suites
Both acceptance suites spawn a real `tsc` against a materialized copy of the
carrier, and both carried their own binary path, flag list, process wrapper and
materialization logic. Two copies of a flag list drift, and a drifted flag set
means the suites measure different compilers — a guard could hold under one and
not the other with nothing to say so.

One module in the helpers tier now owns all of it, including a single
proof-block marker. Three sites had been keying on proof-block comment text
independently, so a reworded comment broke them one at a time.

The shared marker is the EARLIER of the carrier's two proof blocks. Truncating
more than a given relaxation strictly needs is safe: the fixtures fail on the
carrier's types, not on its proofs, so removing proofs never makes a fixture
compile on its own — whereas keeping a block that asserts the relaxed property
makes the copy fail for the right reason, which reads exactly like the guard
holding.

Also corrects an overclaim in the PR body. Both downstream epics name anchor
028 as their blocker and 028 is not fully closed here, so "unblocks" was too
strong. The precise claim is three of the four teeth #1825 enumerates, with the
fourth resting on the premise that turned out false.
